feat(ui/sidebar): add logo for ce
