+++ app/react/sidebar/Footer/UpdateNotifications.module.css
