+++ app/react/sidebar/Footer/UpdateNotifications.tsx
